move feature components to Feature
