+++ src/pages/features/modeling.jsx
@@ -5,8 +5,8 @@ import { withI18n, Trans } from '@lingui/react';
 import { graphql } from 'gatsby';
 import Img from 'gatsby-image';
 
-import { FeatureLinks } from '../../components/Feature';
-import { Title, FeatureList, FeatureLi, Hr } from '../../layouts/utils';
+import { FeatureLinks, FeatureList, FeatureLi } from '../../components/Feature';
+import { Title, Hr } from '../../layouts/utils';
 
 export default withI18n()(({ i18n, ...props }: Props) => (
   <div>
